[D0001] — Hardware build — Workspace established
PS (Mas) - Dev
   ────────────────────────────────────────

The repository has been linked to the TIA Portal
Workspace for the PS2001-pal-software module build.

This build is the TIA Project HARDWARE BUILD only

The CPU is a 1515-2PN (6ES7 515-2AM02-0AB0) SW Ver 2.8

The CPU is designated CON100 and has IP address:

   192.168.1.100

Consequently, the Workspace repository for the TIA Project
now has a corresponding CON100 subfolder this is equipped with:

1. Program blocks
      OB00001_IntNMainProgram.xml

2. PLC tags
      Default tag table.xml

The default tag table, is that produced when the project
was created.

OB1 has been equipped with standard comments and the
MIT Licence. It contains no code at this point.

README.md has been updated for revision information.

   ────────────────────────────────────────
Those aspects of the PAL project that have been migrated to the GitHub
Version Control System (VCS) are operating under the Software Control
Mechanism (SCM) specified in document PS2001-5-2302-011:
   https://practicalseries.com/2001-pal/31-git/11-00-scm.html